Added output folder to construct_event_database and process_output scripts Small fixes to sent_preprocess, utils and extract/ind.py scripts

+++ bin/scripts/to_separate_json_files.py
@@ -1,6 +1,7 @@
 import json
 import sys
 import re
+import os.path
 
 input_json_file = sys.argv[1]
 out_json_folder = sys.argv[2]
@@ -15,7 +16,8 @@ input_file = open(input_json_file, "r", encoding="utf-8")
 for line in input_file:
     d = json.loads(line)
     filename = re.sub(r"(\.ece\d?|\.json|\.html?|\.cms|)$", r".json", d["id"])
-    with open(out_json_folder + "/" + filename, "w", encoding="utf-8") as f:
-        f.write(line)
+    if not os.path.isfile(out_json_folder + "/" + filename):
+        with open(out_json_folder + "/" + filename, "w", encoding="utf-8") as f:
+            f.write(line)
 
 input_file.close()
